// src/types.ts
// Generated by https://quicktype.io

export type ContentType = 'tv' | 'movie';
export const contentLanguage = {
  en: 'en-US',
  fr: 'fr-FR',
};

type SortOrder = 'asc' | 'desc';
type SortField = 'popularity' | 'release' | 'original_title' | 'vote_average' | 'name' | 'title';

export type SortBy = `${SortField}.${SortOrder}`;
export interface Contents {
  page: number;
  results: ContentResult[];
  total_pages: number;
  total_results: number;
}

type ContentResult = Movie & TvShow;

export interface Content {
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  vote_average: number;
  vote_count: number;
}
export interface Movie extends Content {
  adult: boolean;
  title: string;
  video: boolean;
  release_date: string;
}
export interface TvShow extends Content {
  name: string;
  first_air_date: string;
}

// Generated by https://quicktype.io

export interface MovieDetails {
  adult: boolean;
  backdrop_path: string;
  budget: number;
  genres: Genre[];
  homepage: string;
  id: number;
  imdb_id: string;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  revenue: number;
  runtime: number;
  status: string;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  similar: Contents;
}

export interface Genre {
  id: number;
  name: string;
}

// Generated by https://quicktype.io

export interface SerieDetails {
  backdrop_path: string;
  first_air_date: string;
  genres: Genre[];
  homepage: string;
  id: number;
  in_production: boolean;
  last_air_date: string;
  name: string;
  next_episode_to_air: null;
  number_of_episodes: number;
  number_of_seasons: number;
  origin_country: string[];
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  status: string;
  tagline: string;
  type: string;
  vote_average: number;
  vote_count: number;
  similar: Contents;
}